requirements.txt
numpy
torch==1.10.0
torchvision==0.11.1
absl-py==0.11.0
astunparse==1.6.3
boto3==1.17.20
python-decouple
delta
pytube==11.0.2
google_auth_oauthlib
pyspark
keras_ocr
google
botocore==1.20.20
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colormath==3.0.0
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
easydict==1.9
easyocr==1.2.5
fastapi==0.63.0
gast
google-auth==1.27.0
google-pasta==0.2.0
grpcio
h11==0.12.0
httptools==0.1.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.0
imutils==0.5.4
jmespath==0.10.0
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5
oauthlib==3.1.0
opencv-python
opt-einsum==3.3.0
pika==1.2.0
Pillow==8.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8
pygame
pytorchyolo
pymongo==3.12
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.4
scikit-image==0.17.2
scikit-learn==0.24.1
scipy
six==1.15.0
starlette==0.13.6
tensorboard
termcolor==1.1.0
torchaudio
tf-estimator-nightly
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
uvloop==0.15.1
watchgod==0.7
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
kafka-python
pytorch-lightning
lightning-flash==0.5.2
ffmpeg
kornia==0.6.2
webcolors
loguru
pycocotools
thop
keras
tensorflow-gpu
#tensorflow==2.7
you-get==0.4.1555
seaborn
opencv-python
h5py
gpuinfo
scipy
consoleprinter
efficientnet_pytorch
pyflann-py3
mediapipe
